Statement-2-Lucknow-Foodie/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
chromadb
pydantic
python-dotenv
